MAINT: Moved more functions to controller.py
