+++ topobank/manager/models.py
@@ -300,8 +300,8 @@ class Surface(models.Model, SubjectMixin):
         # Request all standard analyses to be available for that user
         #
         _log.info(f"After sharing surface {self.id} with user {with_user.id}, requesting all standard analyses...")
-        from topobank.analysis.models import AnalysisFunction
-        from topobank.analysis.utils import request_analysis
+        from ..analysis.models import AnalysisFunction
+        from ..analysis.controller import request_analysis
         analysis_funcs = AnalysisFunction.objects.all()
         for topo in self.topography_set.all():
             for af in analysis_funcs:
@@ -501,7 +501,7 @@ class Surface(models.Model, SubjectMixin):
         - with this surfaces as subject
         This is done in that order.
         """
-        from ..analysis.utils import renew_analyses_for_subject
+        from ..analysis.controller import renew_analyses_for_subject
 
         if include_topographies:
             _log.info(f"Regenerating analyses of topographies of surface {self.pk}..")
@@ -920,7 +920,7 @@ class Topography(models.Model, SubjectMixin):
 
     def renew_analyses(self):
         """Submit all analysis for this topography."""
-        from ..analysis.utils import renew_analyses_for_subject
+        from ..analysis.controller import renew_analyses_for_subject
         renew_analyses_for_subject(self)
 
     def to_dict(self):
