+++ topobank/manager/tests/test_search.py
@@ -530,6 +530,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
         'folder': False,
         'key': f'topography-{topo1a.pk}',
         'surface_key': f'surface-{surface1.pk}',
+        'label': topo1a.label,
         'name': topo1a.name,
         'id': topo1a.pk,
         'publication_authors': '',
@@ -554,6 +555,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
         'folder': False,
         'key': f'topography-{topo1b.pk}',
         'surface_key': f'surface-{surface1.pk}',
+        'label': topo1b.label,
         'name': topo1b.name,
         'id': topo1b.pk,
         'publication_authors': '',
@@ -579,6 +581,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
         'folder': False,
         'key': f'topography-{topo2a.pk}',
         'surface_key': f'surface-{surface2.pk}',
+        'label': topo2a.label,
         'name': topo2a.name,
         'id': topo2a.pk,
         'publication_authors': '',
@@ -604,6 +607,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
         'folder': False,
         'key': f'topography-{topo2b.pk}',
         'surface_key': f'surface-{surface2.pk}',
+        'label': topo2b.label,
         'name': topo2b.name,
         'id': topo2b.pk,
         'publication_authors': '',
@@ -631,6 +635,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
         'description': '',
         'folder': True,
         'key': f'surface-{surface1.pk}',
+        'label': surface1.label,
         'name': surface1.name,
         'id': surface1.pk,
         'publication_authors': '',
@@ -677,6 +682,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
             'id': bike_pk,
             'key': f'tag-{bike_pk}',
             'folder': True,
+            'label': 'bike',
             'name': 'bike',
             'selected': False,
             'children': [
@@ -698,6 +704,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
             'id': train_pk,
             'key': f"tag-{train_pk}",
             'folder': True,
+            'label': 'train',
             'name': 'train',
             'selected': False,
             'children': [
@@ -724,6 +731,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
                             'publication_date': '',
                             'key': f"tag-{train_ice_restaurant_pk}",
                             'folder': True,
+                            'label': 'train/ice/restaurant',
                             'name': 'train/ice/restaurant',
                             'selected': False,
                             'children': [
@@ -750,6 +758,7 @@ def test_tag_search_with_request_factory(user_three_surfaces_four_topographies):
                     'publication_date': '',
                     'key': f"tag-{train_tgv_pk}",
                     'folder': True,
+                    'label': 'train/tgv',
                     'name': 'train/tgv',
                     'selected': False,
                     'children': [
